Created Forecast with use for getting forecast with a query
