+++ src/models/Forecast/UseCases.ts
@@ -1,0 +1,8 @@
+import { IForecast } from "@interfaces/Forecast";
+import { getForecastData } from "./Services";
+
+const days = +process.env.TOTAL_FORECAST_DAYS;
+
+export const getForecastNextDays = (
+  query: string | number
+): Promise<IForecast> => getForecastData(query, days);
